Exclude checksum from returned value for WifiOpenEVSE
The current regex used to extract the result from the Wifi-returne
HTML includes the RAPI checksum in the value.  This confuses many of
the functions that try to parse the response.  For example, when I
call accumulated_wh(), something like 5920569^1C is extracted.  Then,
the accumulated_wh() function tries to interpret this string as an
int, which raises a ValueError.

This small tweak to the regex moves the checksum out of the match
group, causing it not to be part of the string returned from the
request.