Add support for password authentication.
This change adds support for (optional) use of password
authentication.  To use it, the caller needs to provide a username and
password as arguments when creating the WifiOpenEVSE object.

Note that this doesn't really make your communications with the
OpenEVSE secure in any real sense, since the communication is still
via unencrypted HTTP.  To make it really secure you'd need SSL
support, which I haven't looked into.  Still, it's better than
nothing.